requirements.txt
asgiref==3.4.1
coverage==6.1.2
Django==3.2.9
django-bootstrap-modal-forms==2.2.0
django-environ==0.8.1
pytz==2021.3
sqlparse==0.4.2
